msp/monitoring: split alert builders into many files
